% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reducedSet}
\alias{reducedSet}
\title{Smaller Subset of The Cancer Genome Atlas Cervical Cancer HTSeq Data.}
\format{
The format is of formal class 'DESeqTransform' (package "DESeq2") with the following extractor functions:
\describe{
\item{assay}{includes rld transformed transcript expression data}
\item{colData}{includes phenotypic data: \code{age_at_index}, \code{cigarettes_per_day}, \code{race}, and \code{Stage}}
}
}
\source{
TCGAbiolinks TCGA-CESC
}
\usage{
reducedSet
}
\description{
This subset includes 41 transcripts and four phenotypic variables for
242 subjects with squamous cell cervical cancer.
}
\examples{
\donttest{
library(DESeq2)
data(reducedSet)
head(colData(reducedSet))
}
}
\keyword{datasets}
